--- packages/server/src/graphql/middlewares/authentications.spec.js
@@ -1,7 +1,0 @@
-const index = require('./authentications');
-
-describe('Authentications Middleware', () => {
-  it('returns an object', () => {
-    expect(index).toBeInstanceOf(Object);
-  });
-});
