--- packages/server/src/graphql/middlewares/middlewares.spec.js
@@ -1,7 +1,0 @@
-const index = require('./middlewares');
-
-describe('Middlewares Module', () => {
-  it('index is an object', () => {
-    expect(index).toBeInstanceOf(Object);
-  });
-});
