+++ packages/server/src/graphql/middlewares/notebooks.js
@@ -1,7 +1,5 @@
 const _ = require('lodash');
 
-const { notebookDoesNotExists } = require('../../../config/lang');
-
 const { NotebooksData } = require('../../data');
 const logger = require('../../lib/log');
 
@@ -10,27 +8,6 @@ const { Query } = require('../../data/classes');
 const { Options } = require('../../data/classes');
 
 const log = logger.module('middlewares/notebooks');
-
-/**
- * @param {*} root
- * @param {object} args
- * @param {Defs.NotebookID} args.notebookId
- * @param {Schema.Context} _context
- * @param {object} _info
- * @return {Promise<boolean>}
- */
-async function notebookExistsValidation(root, args, _context, _info) {
-  const notebookId = _.get(args, 'notebookId', null);
-
-  const params = new Params({ id: notebookId });
-  const query = new Query();
-  const options = new Options({ raw: true });
-  const count = await new NotebooksData().getCount(params, query, options);
-
-  if (!count) throw new Error(notebookDoesNotExists());
-
-  return true;
-}
 
 /**
  * @param {object | null} root
@@ -59,7 +36,5 @@ async function setContextNotebook(root, args, context, _info) {
 }
 
 module.exports = {
-  notebookExistsValidation,
-
   setContextNotebook,
 };
