--- packages/server/src/graphql/middlewares/validators.spec.js
@@ -1,7 +1,0 @@
-const index = require('./validators');
-
-describe('Validators Module', () => {
-  it('index is an object', () => {
-    expect(index).toBeInstanceOf(Object);
-  });
-});
