+++ packages/server/src/graphql/schema.js
@@ -8,7 +8,7 @@ const { stats } = require('./types/stats');
 
 const { permissions } = require('./authorizers');
 const { middlewares } = require('./middlewares');
-const { validators } = require('./middlewares');
+const { validators } = require('./validators');
 
 const schema = new graphql.GraphQLSchema({
   types: [
